+++ aci/utils.go
@@ -2,7 +2,6 @@ package aci
 
 import (
 	"fmt"
-	"regexp"
 	"strings"
 
 	"github.com/ciscoecosystem/aci-go-client/container"
@@ -63,8 +62,16 @@ func GetMOName(dn string) string {
 }
 
 func GetParentDn(childDn string) string {
-	re := regexp.MustCompile("(.*)/\\S+-\\S+.*$")
-	match := re.FindStringSubmatch(childDn)
-	return match[1]
+	arr := strings.Split(childDn, "/")
+	// in case of cidr blocks we have extra / in the ip range so let's catch it and remove. This will have extra part.
+	if strings.Contains(childDn, "]") && strings.Contains(childDn, "[") {
+		dnWithRn := strings.Split(childDn, "[")
+
+		slashedArr := strings.Split(dnWithRn[0], "/")
+		return strings.Join(slashedArr[:len(slashedArr)-1], "/")
+
+	}
+
+	return strings.Join(arr[:len(arr)-1], "/")
 
 }